requirements.txt
wheel==0.45.1
flask
mysql-connector-python
pdfminer.six
python-docx
werkzeug
gunicorn
